+++ lib/typings/yargs-parser-types.ts
@@ -51,6 +51,8 @@ export interface Configuration {
     'nargs-eats-options': boolean;
     /** The prefix to use for negated boolean variables. Default is `'no-'` */
     'negation-prefix': string;
+    /** Should positional values that look like numbers be parsed? Default is `true` */
+    'parse-positional-numbers': boolean;
     /** Should keys that look like numbers be treated as such? Default is `true` */
     'parse-numbers': boolean;
     /** Should unparsed flags be stored in -- or _? Default is `false` */
@@ -132,6 +134,7 @@ export interface Parser {
     detailed(args: ArgsInput, opts?: Partial<Options>): DetailedArguments;
     camelCase(str: string): string;
     decamelize(str: string, joinString?: string): string;
+    looksLikeNumber(x: null | undefined | number | string): boolean;
 }
 export declare type StringFlag = Dictionary<string[]>;
 export declare type BooleanFlag = Dictionary<boolean>;